+++ src/constants.h
@@ -45,26 +45,26 @@ BLOCK_ALLOC_FILL_PAGES(callable,2)
 PMOD_EXPORT struct callable *low_make_callable(c_fun fun,
 				   struct pike_string *name,
 				   struct pike_type *type,
-				   INT16 flags,
+				   int flags,
 				   optimize_fun optimize,
 				   docode_fun docode);
 PMOD_EXPORT struct callable *make_callable(c_fun fun,
 			       const char *name,
 			       const char *type,
-			       INT16 flags,
+			       int flags,
 			       optimize_fun optimize,
 			       docode_fun docode);
 PMOD_EXPORT struct callable *add_efun2(const char *name,
 			    c_fun fun,
 			    const char *type,
-			    INT16 flags,
+			    int flags,
 			    optimize_fun optimize,
 			    docode_fun docode);
-PMOD_EXPORT struct callable *add_efun(const char *name, c_fun fun, const char *type, INT16 flags);
+PMOD_EXPORT struct callable *add_efun(const char *name, c_fun fun, const char *type, int flags);
 PMOD_EXPORT struct callable *quick_add_efun(const char *name, ptrdiff_t name_length,
 					    c_fun fun,
 					    const char *type, ptrdiff_t type_length,
-					    INT16 flags,
+					    int flags,
 					    optimize_fun optimize,
 					    docode_fun docode);
 void init_builtin_constants(void);
